Updated firebaseConfig, made folders for pages and moved the respective components into their own page
