+++ src/firebaseConfig.js
@@ -1,12 +1,11 @@
 var firebaseConfig = {
-    //TODO Replace with own config file
-    // apiKey: "AIzaSyD7KhDYomFFrQE4dHH6DDs-SNK9rfptZ9o",
-    // authDomain: "assessment-app-41d0b.firebaseapp.com",
-    // databaseURL: "https://assessment-app-41d0b.firebaseio.com",
-    // projectId: "assessment-app-41d0b",
-    // storageBucket: "assessment-app-41d0b.appspot.com",
-    // messagingSenderId: "418813227371",
-    // appId: "1:418813227371:web:22828bd50e32150598dc2e"
+  apiKey: "AIzaSyD7KhDYomFFrQE4dHH6DDs-SNK9rfptZ9o",
+  authDomain: "assessment-app-41d0b.firebaseapp.com",
+  databaseURL: "https://assessment-app-41d0b.firebaseio.com",
+  projectId: "assessment-app-41d0b",
+  storageBucket: "assessment-app-41d0b.appspot.com",
+  messagingSenderId: "418813227371",
+  appId: "1:418813227371:web:22828bd50e32150598dc2e",
 };
 
 export default firebaseConfig;
